+++ AWSAuthSDK/Tests/AWSMobileClientTests/AWSMobileClientTests.swift
@@ -241,6 +241,50 @@ class AWSMobileClientTests: XCTestCase {
         signIn(username: username)
     }
     
+    /// Test successful sign in and order of callback
+    ///
+    /// - Given: An unauthenticated session
+    /// - When:
+    ///    - I add a user state listener
+    ///    - I invoke `signIn` with a completion callback
+    /// - Then:
+    ///    - The user state is `signedIn`
+    ///    - My user state listener callback is invoked first
+    ///    - My `signIn` completion callback is invoked last
+    ///
+    func testOrderOfCallbackInSuccessfulSignIn() {
+        let username = "testUser" + UUID().uuidString
+        let signInListenerWasSuccessful = expectation(description: "signIn listener was successful")
+        let signInWasSuccessful = expectation(description: "signIn was successful")
+        signUpAndVerifyUser(username: username)
+
+        AWSMobileClient.sharedInstance().addUserStateListener(self) { (userState, info) in
+            switch (userState) {
+            case .signedIn:
+                signInListenerWasSuccessful.fulfill()
+                print("Listener user is signed in.")
+            default:
+                print("Listener \(userState)")
+            }
+        }
+
+        AWSMobileClient.sharedInstance().signIn(username: username, password: sharedPassword) { (signInResult, error) in
+            if let error = error {
+                XCTFail("User login failed: \(error.localizedDescription)")
+                return
+            }
+
+            guard let signInResult = signInResult else {
+                XCTFail("User login failed, signInResult unexpectedly nil")
+                return
+            }
+            XCTAssertEqual(signInResult.signInState, .signedIn, "Could not verify sign in state")
+            signInWasSuccessful.fulfill()
+        }
+        wait(for: [signInListenerWasSuccessful, signInWasSuccessful], timeout: 5, enforceOrder: true)
+        AWSMobileClient.sharedInstance().removeUserStateListener(self)
+    }
+
     func testSignInFailCase() {
         let username = "testUser" + UUID().uuidString
         signUpAndVerifyUser(username: username)
@@ -622,6 +666,64 @@ class AWSMobileClientTests: XCTestCase {
         XCTAssertTrue(AWSMobileClient.sharedInstance().isSignedIn == false, "Expected to return false for isSignedIn")
     }
     
+    /// Test successful sign out and order of callback
+    ///
+    /// - Given: An unauthenticated session
+    /// - When:
+    ///    - I add a user state listener
+    ///    - I invoke `signIn` with a completion callback
+    ///    - I invoke `signOut` with a completion callback
+    /// - Then:
+    ///    - The user state is `signedOut`
+    ///    - My user state listener callback is invoked first with signedIn
+    ///    - My user state listener callback is invoked first with signedOut
+    ///    - My `signOut` completion callback is invoked last
+    ///
+    func testOrderOfCallbacksForSignOut() {
+        let username = "testUser" + UUID().uuidString
+        let signoutExpectation = expectation(description: "Successfully signout")
+        let signInListenerExpectation = expectation(description: "signIn was successful")
+        let signOutListenerExpectation = expectation(description: "signOut was successful")
+        signUpAndVerifyUser(username: username)
+        XCTAssertEqual(AWSMobileClient.sharedInstance().currentUserState, .signedOut)
+        
+        AWSMobileClient.sharedInstance().addUserStateListener(self) { (userState, info) in
+            switch (userState) {
+            case .signedIn:
+                signInListenerExpectation.fulfill()
+                print("Listener user is signed in.")
+            case .signedOut:
+                signOutListenerExpectation.fulfill()
+                print("Listener user signed out.")
+            default:
+                print("Listener \(userState)")
+            }
+        }
+        signIn(username: username)
+        XCTAssertTrue(AWSMobileClient.sharedInstance().isSignedIn == true, "Expected to return true for isSignedIn")
+        sleep(1)
+        AWSMobileClient.sharedInstance().signOut { (error) in
+            XCTAssertTrue(AWSMobileClient.sharedInstance().isSignedIn == false, "Expected to return false for isSignedIn")
+            signoutExpectation.fulfill()
+        }
+        wait(for: [signInListenerExpectation, signOutListenerExpectation, signoutExpectation], timeout: 2, enforceOrder: true)
+        AWSMobileClient.sharedInstance().removeUserStateListener(self)
+    }
+    
+    func testSignOutWithCallback() {
+        let username = "testUser" + UUID().uuidString
+        let signoutExpectation = expectation(description: "Successfully signout")
+        signUpAndVerifyUser(username: username)
+        signIn(username: username)
+        XCTAssertTrue(AWSMobileClient.sharedInstance().isSignedIn == true, "Expected to return true for isSignedIn")
+        sleep(1)
+        AWSMobileClient.sharedInstance().signOut { (error) in
+            XCTAssertTrue(AWSMobileClient.sharedInstance().isSignedIn == false, "Expected to return false for isSignedIn")
+            signoutExpectation.fulfill()
+        }
+        wait(for: [signoutExpectation], timeout: 2)
+    }
+
     func testFederatedSignInDeveloperAuthenticatedIdentities() {
         let getOpendIdRequest = AWSCognitoIdentityGetOpenIdTokenForDeveloperIdentityInput()
         getOpendIdRequest?.identityPoolId = identityPoolId